[feature/improve_jetpack_navigation] Update NavGraph and app style
