+++ feature/root/src/main/java/com/alxnophis/jetpack/root/ui/RootActivity.kt
@@ -3,8 +3,6 @@ package com.alxnophis.jetpack.root.ui
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
 import com.alxnophis.jetpack.root.di.injectRoot
 import com.alxnophis.jetpack.root.navigation.SetupNavGraph
 import com.alxnophis.jetpack.root.ui.viewmodel.RootViewModel
@@ -13,7 +11,6 @@ import org.koin.androidx.viewmodel.ext.android.viewModel
 class RootActivity : ComponentActivity() {
 
     private val viewModel: RootViewModel by viewModel()
-    private lateinit var navController: NavHostController
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -23,8 +20,7 @@ class RootActivity : ComponentActivity() {
 
     private fun renderContent() {
         setContent {
-            navController = rememberNavController()
-            SetupNavGraph(navController)
+            SetupNavGraph()
         }
     }
 }
