+++ shared/core/src/main/java/com/alxnophis/jetpack/core/ui/theme/Palette.kt
@@ -1,0 +1,18 @@
+package com.alxnophis.jetpack.core.ui.theme
+
+import androidx.compose.material.lightColors
+
+internal val lightColorPalette = lightColors(
+    primary = coreColorGreen,
+    primaryVariant = coreColorGreenDark,
+    onPrimary = coreColorWhite,
+    secondary = coreColorOrange,
+    secondaryVariant = coreColorOrangeDark,
+    onSecondary = coreColorVulcan,
+    error = coreColorRed,
+    onError = coreColorWhite,
+    surface = coreColorWhite,
+    onSurface = coreColorCarbon,
+    background = coreColorWhite,
+    onBackground = coreColorCarbon,
+)
