+++ shared/core/src/main/res/values/styles.xml
@@ -1,9 +1,4 @@
 <resources>
 
-    <style name="CoreTheme" parent="Theme.MaterialComponents.DayNight.NoActionBar">
-        <item name="android:windowBackground">#FFFFFF</item>
-        <item name="colorPrimary">#26B378</item>
-        <item name="colorPrimaryDark">#1E8F60</item>
-    </style>
-
+    <style name="CoreTheme" parent="Theme.MaterialComponents.DayNight.NoActionBar" />
 </resources>
